requirement.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
blis==0.4.1
bs4==0.0.1
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
confusable-homoglyphs==3.2.0
cryptography==3.2.1
dataclasses==0.6
django-allauth==0.44.0
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-extra-views==0.13.0
django-humanize==0.1.2
django-registration==3.1.1
django-rest-auth==0.9.5
django-unixtimestampfield==0.4.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
Faker==4.18.0
humanize==3.1.0
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
joblib==0.17.0
localStoragePy==0.1.7
oauthlib==3.1.0
olefile==0.46
packaging==20.4
PyDictionary==2.0.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-resize-image==1.1.19
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
urllib3==1.26.2
wasabi==0.8.0
